Migrate IDs to INT

+++ api/src/main/java/com/codeforcommunity/dto/school/SchoolSummary.java
@@ -4,21 +4,21 @@ import com.codeforcommunity.enums.Country;
 
 public class SchoolSummary {
 
-  private Long id;
+  private Integer id;
   private String name;
   private Country country;
 
-  public SchoolSummary(Long id, String name, Country country) {
+  public SchoolSummary(Integer id, String name, Country country) {
     this.id = id;
     this.name = name;
     this.country = country;
   }
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
